Implement functionality for AA assignment of JS and CM
